Relocate 21 zero-column tags from components/game_state.h to tags/tags.h
Per the single-tags-header convention and Fabian drift issue #1276.

The audit listed 17 tags; this PR moves all 21 zero-column structs found
in app/components/game_state.h (the original 17 phase / pending-transition
/ level-select-confirmation tags plus EndChoiceRestart / EndChoiceLevelSelect
/ EndChoiceMainMenu / EnergyDepletedDeath — same Fabian principle, same
single-header convention).

After the move, app/components/game_state.h carries only data components
(GameState{phase_timer}, LevelSelectState, TerminalResultState).

Back-compat: game_state.h #include's tags/tags.h so existing consumers
(~60 .cpp/.h files) that pulled these tags in via game_state.h still resolve
without per-site changes. A follow-up cycle can opt to thin out the
back-compat include site-by-site (or replace with direct tags/tags.h
includes) once the affected systems are touched for other reasons.

Build: zero warnings (-Wall -Wextra -Werror, AppleClang).
Tests: 893 passed / 0 failed (same as baseline; 1 smoke test skipped on
host as expected).

Closes #1276.
